# docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /v1/beer/brands:
    get:
      tags:
      - beer-controller
      operationId: getBrandsList
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
    post:
      tags:
      - beer-controller
      operationId: saveBrand
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewBeerBrand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Brand"
  /v1/beer/brands/{brandId}/flavours:
    get:
      tags:
      - beer-controller
      operationId: getBrandNames
      parameters:
      - name: brandId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BeerFlavour"
    post:
      tags:
      - beer-controller
      operationId: saveBrandName
      parameters:
      - name: brandId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewBeerFlavour"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BeerFlavour"
...
  /v1/recommendations/edit:
    patch:
      tags:
      - recommendations-controller
      operationId: reorderRecommendations
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/RecommendationUpdate"
        required: true
      responses:
        "200":
          description: Updated recommendations with new order numbers and names
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
  /v1/recommendations/{id}:
    delete:
      tags:
      - recommendations-controller
      operationId: deleteRecommendation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Recommendation deleted
...
  /v1/beer/consumption-types:
    get:
      tags:
      - beer-controller
      operationId: getConsumptionTypesList
      parameters:
      - name: amount
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsumptionType"
...
  /v1/admin/recommendations/list:
    get:
      tags:
      - admin-recommendations-controller
      operationId: getRecommendationsList_1
      description: Admin only. The default recommendations every user starts with, sorted by orderNumber.
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DefaultRecommendation"
        "403":
          description: caller is not in the /admin group
  /v1/admin/recommendations:
    post:
      tags:
      - admin-recommendations-controller
      operationId: createDefaultRecommendation
      description: Admin only. The new recommendation is placed after the current last one.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewDefaultRecommendation"
        required: true
      responses:
        "200":
          description: The saved recommendation
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DefaultRecommendation"
        "400":
          description: name, alcoholTypeId, colorPaletteId or glasswareId is missing
        "403":
          description: caller is not in the /admin group
  /v1/admin/recommendations/edit:
    patch:
      tags:
      - admin-recommendations-controller
      operationId: reorderRecommendations_1
      description: >-
        Admin only. Renumbers the listed recommendations 1..N in request order and
        sets their names. Send the full list: rows left out keep their old order number.
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/RecommendationUpdate"
        required: true
      responses:
        "200":
          description: All default recommendations, sorted by orderNumber
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DefaultRecommendation"
        "400":
          description: an entry is missing id or name
        "403":
          description: caller is not in the /admin group
  /v1/admin/recommendations/{id}:
    delete:
      tags:
      - admin-recommendations-controller
      operationId: deleteRecommendation_1
      description: Admin only.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Recommendation deleted
        "403":
          description: caller is not in the /admin group
        "404":
          description: no default recommendation with that id
...
  /v1/admin/alcohol/types/{id}:
    patch:
      tags:
      - admin-alcohol-controller
      operationId: patchAlcoholType
      description: Admin only. Applies the supplied non-null fields.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlcoholType"
      responses:
        "200":
          description: Updated alcohol type
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlcoholType"
        "400":
          description: name exceeds 100 characters or volumeIds contains more than 50 entries
        "403":
          description: caller is not in the /admin group
        "404":
          description: no alcohol type with that id
    delete:
      tags:
      - admin-alcohol-controller
      operationId: deleteAlcoholType
      description: Admin only. Deletes the alcohol type.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Alcohol type deleted
        "403":
          description: caller is not in the /admin group
        "404":
          description: no alcohol type with that id
        "409":
          description: alcohol type is referenced by other records
  /v1/admin/alcohol/subtypes/{id}:
    patch:
      tags:
      - admin-alcohol-controller
      operationId: patchAlcoholSubtypeType
      description: Admin only. Applies the supplied non-null fields.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlcoholSubtype"
      responses:
        "200":
          description: Updated alcohol subtype
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AlcoholSubtype"
        "403":
          description: caller is not in the /admin group
        "404":
          description: no alcohol subtype with that id
    delete:
      tags:
      - admin-alcohol-controller
      operationId: deleteAlcoholSubtypeType
      description: Admin only. Deletes the alcohol subtype.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Alcohol subtype deleted
        "403":
          description: caller is not in the /admin group
        "404":
          description: no alcohol subtype with that id
        "409":
          description: alcohol subtype is referenced by other records
...
  /v1/admin/default/beer/brands:
    post:
      tags:
      - admin-beer-default-controller
      operationId: createDefaultBeerBrand
      description: >-
        Admin only. Creates a default beer brand and optional flavours owned by the configured admin UUID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewBeerBrand"
      responses:
        "200":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
        "400":
          description: Invalid request body or parameters
        "401":
          description: Authentication required
        "403":
          description: caller is not in the /admin group
    get:
      tags:
      - admin-beer-default-controller
      operationId: getDefaultBeerBrands
      description: Admin only. Lists default beer brands.
      responses:
        "200":
          description: Default beer brands
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
        "403":
          description: caller is not in the /admin group
  /v1/admin/default/beer/brands/{brandId}/flavours:
    post:
      tags:
      - admin-beer-default-controller
      operationId: createDefaultBeerFlavour
      description: >-
        Admin only. Creates a default beer flavour for the path brandId, owned by the configured admin UUID.
      parameters:
      - name: brandId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewBeerFlavour"
      responses:
        "200":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeerFlavour"
        "400":
          description: Invalid request body or parameters
        "401":
          description: Authentication required
        "403":
          description: caller is not in the /admin group
    get:
      tags:
      - admin-beer-default-controller
      operationId: getDefaultBeerFlavours
      description: Admin only. Lists default flavours for a beer brand.
      parameters:
      - name: brandId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Default beer flavours
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BeerFlavour"
        "403":
          description: caller is not in the /admin group
  /v1/admin/default/beer/consumption-types:
    post:
      tags:
      - admin-beer-default-controller
      operationId: createDefaultConsumptionType
      description: >-
        Admin only. Creates a shared beer consumption type with a name and glassware ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewConsumptionType"
      responses:
        "200":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumptionType"
        "400":
          description: Invalid request body or parameters
        "401":
          description: Authentication required
        "403":
          description: caller is not in the /admin group
    get:
      tags:
      - admin-beer-default-controller
      operationId: getDefaultBeerConsumptionTypes
      description: Admin only. Lists all default beer consumption types in name order.
      responses:
        "200":
          description: Default beer consumption types
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsumptionType"
        "403":
          description: caller is not in the /admin group
...
  /v1/admin/user-defined/beer/brands:
    get:
      tags:
      - admin-beer-user-controller
      operationId: getUserDefinedBeerBrands
      description: Admin only. Lists user-defined beer brands awaiting moderation.
      responses:
        "200":
          description: User-defined beer brands
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
        "403":
          description: caller is not in the /admin group
  /v1/admin/user-defined/beer/brands/{brandId}/flavours:
    get:
      tags:
      - admin-beer-user-controller
      operationId: getUserDefinedBeerFlavours
      description: Admin only. Lists user-defined flavours for a beer brand.
      parameters:
      - name: brandId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: User-defined beer flavours
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BeerFlavour"
        "403":
          description: caller is not in the /admin group
components:
  schemas:
...
    NewBeerBrand:
      type: object
      required: [name, colorPaletteId]
      properties:
        name:
          type: string
          maxLength: 100
        flavours:
          type: array
          items:
            type: string
        colorPaletteId:
          type: integer
          format: int32
    NewConsumptionType:
      type: object
      required: [name, glasswareId]
      properties:
        name:
          type: string
        glasswareId:
          type: integer
          format: int32
...
    Brand:
      type: object
      properties:
        id:
          type: integer
          format: int32
        userId:
          type: string
          format: uuid
        name:
          type: string
        colorPaletteId:
          type: integer
          format: int32
    NewBeerFlavour:
      type: object
      required: [name, colorPaletteId]
      properties:
        name:
          type: string
          maxLength: 100
        colorPaletteId:
          type: integer
          format: int32
    BeerFlavour:
      type: object
      properties:
        id:
          type: integer
          format: int32
        brandId:
          type: integer
          format: int32
        userId:
          type: string
          format: uuid
        name:
          type: string
        colorPaletteId:
          type: integer
          format: int32
...
    AlcoholType:
      type: object
      properties:
        id:
          type: integer
          format: int32
        userId:
          type: string
          format: uuid
        name:
          type: string
        volumeIds:
          type: array
          items:
            type: integer
            format: int32
        colorPaletteId:
          type: integer
          format: int32
        glasswareId:
          type: integer
          format: int32
    UpdateAlcoholType:
      type: object
      description: Omitted or null fields are left unchanged.
      properties:
        name:
          type: [string, "null"]
          maxLength: 100
        volumeIds:
          type: [array, "null"]
          maxItems: 50
          items:
            type: integer
            format: int32
        colorPaletteId:
          type: [integer, "null"]
          format: int32
        glasswareId:
          type: [integer, "null"]
          format: int32
...
    AlcoholSubtype:
      type: object
      properties:
        id:
          type: integer
          format: int64
        alcoholTypeId:
          type: integer
          format: int32
        userId:
          type: string
          format: uuid
        name:
          type: string
        colorPaletteId:
          type: integer
          format: int32
        glasswareId:
          type: integer
          format: int32
    UpdateAlcoholSubtype:
      type: object
      description: Omitted or null fields are left unchanged.
      properties:
        name:
          type: [string, "null"]
        colorPaletteId:
          type: [integer, "null"]
          format: int32
        glasswareId:
          type: [integer, "null"]
          format: int32
    Recommendation:
      type: object
      properties:
        id:
          type: integer
          format: int32
        userId:
          type: string
          format: uuid
        name:
          type: string
        alcoholTypeId:
          type: integer
          format: int32
        alcoholSubtypeId:
          type: integer
          format: int64
        alcoholVolumeId:
          type: integer
          format: int32
        brandId:
          type: integer
          format: int32
        beerFlavourId:
          type: integer
          format: int32
        consumptionTypeId:
          type: integer
          format: int32
        endDate:
          type: string
          format: date-time
        colorPaletteId:
          type: integer
          format: int32
        glasswareId:
          type: integer
          format: int32
        orderNumber:
          type: integer
          format: int32
    RecommendationUpdate:
      type: object
      required:
      - id
      - name
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
...
    ConsumptionType:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        glasswareId:
          type: integer
          format: int32
...
    DefaultRecommendation:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        alcoholTypeId:
          type: integer
          format: int32
        alcoholSubtypeId:
          type: integer
          format: int64
        alcoholVolumeId:
          type: integer
          format: int32
        brandId:
          type: integer
          format: int32
        beerFlavourId:
          type: integer
          format: int32
        consumptionTypeId:
          type: integer
          format: int32
        colorPaletteId:
          type: integer
          format: int32
        glasswareId:
          type: integer
          format: int32
        orderNumber:
          type: integer
          format: int32
    NewDefaultRecommendation:
      type: object
      required:
      - name
      - alcoholTypeId
      - colorPaletteId
      - glasswareId
      properties:
        name:
          type: string
        alcoholTypeId:
          type: integer
          format: int32
        alcoholSubtypeId:
          type: integer
          format: int64
        alcoholVolumeId:
          type: integer
          format: int32
        brandId:
          type: integer
          format: int32
        beerFlavourId:
          type: integer
          format: int32
        consumptionTypeId:
          type: integer
          format: int32
        colorPaletteId:
          type: integer
          format: int32
        glasswareId:
          type: integer
          format: int32